Move llama_model_quantize() into llama.cpp
